client/requirements.txt
psutil==5.9.8
websocket-client==1.7.0
pywin32==306
requests==2.31.0
